# scimext/internal/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn3AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn7AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
		"PATCH":  "Authorization,Content-Type",
		"PUT":    "Authorization,Content-Type",
	}
	rn1AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn5AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
		"PATCH":  "Authorization,Content-Type",
		"PUT":    "Authorization,Content-Type",
	}
	rn9AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'A': // Prefix: "Agent"

				if l := len("Agent"); len(elem) >= l && elem[0:l] == "Agent" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'i': // Prefix: "icApplications"

					if l := len("icApplications"); len(elem) >= l && elem[0:l] == "icApplications" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleListAgenticApplicationsRequest([0]string{}, elemIsEscaped, w, r)
						case "POST":
							s.handleCreateAgenticApplicationRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET,POST",
								allowedHeaders: rn3AllowedHeaders,
								acceptPost:     "application/scim+json",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleDeleteAgenticApplicationRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "GET":
								s.handleGetAgenticApplicationRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "PATCH":
								s.handlePatchAgenticApplicationRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "PUT":
								s.handleReplaceAgenticApplicationRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "DELETE,GET,PATCH,PUT",
									allowedHeaders: rn7AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "application/scim+json",
								})
							}

							return
						}

					}

				case 's': // Prefix: "s"

					if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleListAgentsRequest([0]string{}, elemIsEscaped, w, r)
						case "POST":
							s.handleCreateAgentRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET,POST",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "application/scim+json",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleDeleteAgentRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "GET":
								s.handleGetAgentRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "PATCH":
								s.handlePatchAgentRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "PUT":
								s.handleReplaceAgentRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "DELETE,GET,PATCH,PUT",
									allowedHeaders: rn5AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "application/scim+json",
								})
							}

							return
						}

					}

				}

			case 'S': // Prefix: "ServiceProviderConfig"

				if l := len("ServiceProviderConfig"); len(elem) >= l && elem[0:l] == "ServiceProviderConfig" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetServiceProviderConfigRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn9AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'A': // Prefix: "Agent"

				if l := len("Agent"); len(elem) >= l && elem[0:l] == "Agent" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'i': // Prefix: "icApplications"

					if l := len("icApplications"); len(elem) >= l && elem[0:l] == "icApplications" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = ListAgenticApplicationsOperation
							r.summary = "List agentic applications"
							r.operationID = "listAgenticApplications"
							r.operationGroup = ""
							r.pathPattern = "/AgenticApplications"
							r.args = args
							r.count = 0
							return r, true
						case "POST":
							r.name = CreateAgenticApplicationOperation
							r.summary = "Create an agentic application"
							r.operationID = "createAgenticApplication"
							r.operationGroup = ""
							r.pathPattern = "/AgenticApplications"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "DELETE":
								r.name = DeleteAgenticApplicationOperation
								r.summary = "Delete an agentic application"
								r.operationID = "deleteAgenticApplication"
								r.operationGroup = ""
								r.pathPattern = "/AgenticApplications/{id}"
								r.args = args
								r.count = 1
								return r, true
							case "GET":
								r.name = GetAgenticApplicationOperation
								r.summary = "Get an agentic application by ID"
								r.operationID = "getAgenticApplication"
								r.operationGroup = ""
								r.pathPattern = "/AgenticApplications/{id}"
								r.args = args
								r.count = 1
								return r, true
							case "PATCH":
								r.name = PatchAgenticApplicationOperation
								r.summary = "Patch an agentic application"
								r.operationID = "patchAgenticApplication"
								r.operationGroup = ""
								r.pathPattern = "/AgenticApplications/{id}"
								r.args = args
								r.count = 1
								return r, true
							case "PUT":
								r.name = ReplaceAgenticApplicationOperation
								r.summary = "Replace an agentic application"
								r.operationID = "replaceAgenticApplication"
								r.operationGroup = ""
								r.pathPattern = "/AgenticApplications/{id}"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				case 's': // Prefix: "s"

					if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = ListAgentsOperation
							r.summary = "List agents"
							r.operationID = "listAgents"
							r.operationGroup = ""
							r.pathPattern = "/Agents"
							r.args = args
							r.count = 0
							return r, true
						case "POST":
							r.name = CreateAgentOperation
							r.summary = "Create an agent"
							r.operationID = "createAgent"
							r.operationGroup = ""
							r.pathPattern = "/Agents"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "DELETE":
								r.name = DeleteAgentOperation
								r.summary = "Delete an agent"
								r.operationID = "deleteAgent"
								r.operationGroup = ""
								r.pathPattern = "/Agents/{id}"
								r.args = args
								r.count = 1
								return r, true
							case "GET":
								r.name = GetAgentOperation
								r.summary = "Get an agent by ID"
								r.operationID = "getAgent"
								r.operationGroup = ""
								r.pathPattern = "/Agents/{id}"
								r.args = args
								r.count = 1
								return r, true
							case "PATCH":
								r.name = PatchAgentOperation
								r.summary = "Patch an agent"
								r.operationID = "patchAgent"
								r.operationGroup = ""
								r.pathPattern = "/Agents/{id}"
								r.args = args
								r.count = 1
								return r, true
							case "PUT":
								r.name = ReplaceAgentOperation
								r.summary = "Replace an agent"
								r.operationID = "replaceAgent"
								r.operationGroup = ""
								r.pathPattern = "/Agents/{id}"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				}

			case 'S': // Prefix: "ServiceProviderConfig"

				if l := len("ServiceProviderConfig"); len(elem) >= l && elem[0:l] == "ServiceProviderConfig" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetServiceProviderConfigOperation
						r.summary = "Get service provider configuration"
						r.operationID = "getServiceProviderConfig"
						r.operationGroup = ""
						r.pathPattern = "/ServiceProviderConfig"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}