# scimext/openapi/scim-agent-extension.yaml
openapi: 3.0.3
info:
  title: SCIM Agent Extension API
  description: |
    OpenAPI specification for the SCIM Agents and Agentic Applications Extension.

    Based on draft-abbey-scim-agent-extension-00:
    https://datatracker.ietf.org/doc/draft-abbey-scim-agent-extension

    This extension adds two new resource types to SCIM 2.0:
    - Agent: Represents AI agents with their own identity, metadata, and privileges
    - AgenticApplication: Represents applications that host or provide access to agents
  version: 0.1.0
  contact:
    name: AIStandards.io
    url: https://aistandards.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://scim.example.com/v2
    description: Example SCIM server
...
paths:
  /Agents:
    get:
      tags:
        - Agents
      operationId: listAgents
      summary: List agents
      description: |
        Retrieve a list of agents. Supports SCIM filtering on name and externalId.
      parameters:
        - $ref: '#/components/parameters/filter'
        - $ref: '#/components/parameters/startIndex'
        - $ref: '#/components/parameters/count'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/excludedAttributes'
      responses:
        '200':
          description: Successful response
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Agents
      operationId: createAgent
      summary: Create an agent
      description: Create a new agent resource.
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /Agents/{id}:
    parameters:
      - $ref: '#/components/parameters/resourceId'
    get:
      tags:
        - Agents
      operationId: getAgent
      summary: Get an agent by ID
      description: Retrieve a single agent by its unique identifier.
      parameters:
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/excludedAttributes'
      responses:
        '200':
          description: Successful response
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Agents
      operationId: replaceAgent
      summary: Replace an agent
      description: Replace an existing agent resource entirely.
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '200':
          description: Agent replaced successfully
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Agents
      operationId: patchAgent
      summary: Patch an agent
      description: Partially update an existing agent resource using SCIM PATCH operations.
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/PatchRequest'
      responses:
        '200':
          description: Agent patched successfully
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Agents
      operationId: deleteAgent
      summary: Delete an agent
      description: Delete an existing agent resource.
      responses:
        '204':
          description: Agent deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /AgenticApplications:
    get:
      tags:
        - AgenticApplications
      operationId: listAgenticApplications
      summary: List agentic applications
      description: |
        Retrieve a list of agentic applications. Supports SCIM filtering on name, externalId, and applicationUrls.value.
      parameters:
        - $ref: '#/components/parameters/filter'
        - $ref: '#/components/parameters/startIndex'
        - $ref: '#/components/parameters/count'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/excludedAttributes'
      responses:
        '200':
          description: Successful response
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/AgenticApplicationListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - AgenticApplications
      operationId: createAgenticApplication
      summary: Create an agentic application
      description: Create a new agentic application resource.
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/AgenticApplicationCreate'
      responses:
        '201':
          description: Agentic application created successfully
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/AgenticApplication'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /AgenticApplications/{id}:
    parameters:
      - $ref: '#/components/parameters/resourceId'
    get:
      tags:
        - AgenticApplications
      operationId: getAgenticApplication
      summary: Get an agentic application by ID
      description: Retrieve a single agentic application by its unique identifier.
      parameters:
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/excludedAttributes'
      responses:
        '200':
          description: Successful response
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/AgenticApplication'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - AgenticApplications
      operationId: replaceAgenticApplication
      summary: Replace an agentic application
      description: Replace an existing agentic application resource entirely.
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/AgenticApplicationCreate'
      responses:
        '200':
          description: Agentic application replaced successfully
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/AgenticApplication'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - AgenticApplications
      operationId: patchAgenticApplication
      summary: Patch an agentic application
      description: Partially update an existing agentic application resource using SCIM PATCH operations.
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/PatchRequest'
      responses:
        '200':
          description: Agentic application patched successfully
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/AgenticApplication'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - AgenticApplications
      operationId: deleteAgenticApplication
      summary: Delete an agentic application
      description: Delete an existing agentic application resource.
      responses:
        '204':
          description: Agentic application deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /ServiceProviderConfig:
    get:
      tags:
        - ServiceProviderConfig
      operationId: getServiceProviderConfig
      summary: Get service provider configuration
      description: |
        Retrieve the service provider configuration including agent extension support.
      responses:
        '200':
          description: Successful response
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ServiceProviderConfig'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    resourceId:
      name: id
      in: path
      required: true
      description: The unique identifier of the resource.
      schema:
        type: string
    filter:
      name: filter
      in: query
      description: SCIM filter expression.
      schema:
        type: string
    startIndex:
      name: startIndex
      in: query
      description: 1-based index of the first result to return.
      schema:
        type: integer
        minimum: 1
        default: 1
    count:
      name: count
      in: query
      description: Maximum number of results to return.
      schema:
        type: integer
        minimum: 0
    sortBy:
      name: sortBy
      in: query
      description: Attribute to sort by.
      schema:
        type: string
    sortOrder:
      name: sortOrder
      in: query
      description: Sort order (ascending or descending).
      schema:
        type: string
        enum:
          - ascending
          - descending
    attributes:
      name: attributes
      in: query
      description: Comma-separated list of attributes to include in the response.
      schema:
        type: string
    excludedAttributes:
      name: excludedAttributes
      in: query
      description: Comma-separated list of attributes to exclude from the response.
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request - invalid syntax or parameters
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict - resource already exists or version mismatch
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Entitlement:
      type: object
      description: An entitlement granted to an agent.
      properties:
        value:
          type: string
          description: The value of the entitlement.
        display:
          type: string
          description: A human-readable name for the entitlement.
        type:
          type: string
          description: A label indicating the entitlement's function.
        primary:
          type: boolean
          description: Whether this is the primary entitlement.
...
    Agent:
      type: object
      description: |
        An AI agent with its own identifier, metadata, and privileges.
        Schema URI: urn:ietf:params:scim:schemas:core:2.0:Agent
      required:
        - schemas
        - name
      properties:
        schemas:
          type: array
          items:
            type: string
          description: The schemas used for this resource.
          example:
            - "urn:ietf:params:scim:schemas:core:2.0:Agent"
        id:
          type: string
          description: The unique identifier for the agent.
          readOnly: true
        externalId:
          type: string
          description: An external identifier for the agent.
        meta:
          $ref: '#/components/schemas/Meta'
        name:
          type: string
          description: The unique name of the agent. REQUIRED.
        displayName:
          type: string
          description: A human-readable display name for the agent.
        description:
          type: string
          description: A description of the agent.
        agentType:
          type: string
          description: The type of agent (e.g., Assistant, Researcher, Chat bot).
        active:
          type: boolean
          description: Whether the agent is administratively active.
        subject:
          type: string
          description: The subject for inbound token federation correlation.
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupReference'
          description: Groups to which the agent belongs.
          readOnly: true
        entitlements:
          type: array
          items:
            $ref: '#/components/schemas/Entitlement'
          description: Entitlements granted to the agent.
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
          description: Roles assigned to the agent.
        x509Certificates:
          type: array
          items:
            $ref: '#/components/schemas/X509Certificate'
          description: X.509 certificates associated with the agent.
        applications:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationReference'
          description: Agentic applications this agent is associated with.
          readOnly: true
        protocols:
          type: array
          items:
            $ref: '#/components/schemas/Protocol'
          description: Communication protocols supported by this agent.
        parent:
          $ref: '#/components/schemas/Reference'
        owners:
          type: array
          items:
            $ref: '#/components/schemas/Reference'
          description: Users or groups that own this agent.
    AgentCreate:
      type: object
      description: Request body for creating or replacing an agent.
      required:
        - schemas
        - name
      properties:
        schemas:
          type: array
          items:
            type: string
          description: The schemas used for this resource.
          example:
            - "urn:ietf:params:scim:schemas:core:2.0:Agent"
        externalId:
          type: string
          description: An external identifier for the agent.
        name:
          type: string
          description: The unique name of the agent. REQUIRED.
        displayName:
          type: string
          description: A human-readable display name for the agent.
        description:
          type: string
          description: A description of the agent.
        agentType:
          type: string
          description: The type of agent.
        active:
          type: boolean
          description: Whether the agent is administratively active.
        subject:
          type: string
          description: The subject for inbound token federation correlation.
        entitlements:
          type: array
          items:
            $ref: '#/components/schemas/Entitlement'
          description: Entitlements granted to the agent.
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
          description: Roles assigned to the agent.
        x509Certificates:
          type: array
          items:
            $ref: '#/components/schemas/X509Certificate'
          description: X.509 certificates associated with the agent.
        protocols:
          type: array
          items:
            $ref: '#/components/schemas/Protocol'
          description: Communication protocols supported by this agent.
        parent:
          $ref: '#/components/schemas/Reference'
        owners:
          type: array
          items:
            $ref: '#/components/schemas/Reference'
          description: Users or groups that own this agent.
    AgentListResponse:
      type: object
      description: SCIM list response for agents.
      required:
        - schemas
        - totalResults
      properties:
        schemas:
          type: array
          items:
            type: string
          example:
            - "urn:ietf:params:scim:api:messages:2.0:ListResponse"
        totalResults:
          type: integer
          description: Total number of results matching the query.
        startIndex:
          type: integer
          description: 1-based index of the first result returned.
        itemsPerPage:
          type: integer
          description: Number of results returned in this response.
        Resources:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
          description: The list of agents.
...
    PatchRequest:
      type: object
      description: SCIM PATCH request.
      required:
        - schemas
        - Operations
      properties:
        schemas:
          type: array
          items:
            type: string
          example:
            - "urn:ietf:params:scim:api:messages:2.0:PatchOp"
        Operations:
          type: array
          items:
            $ref: '#/components/schemas/PatchOperation'
          description: The list of patch operations to perform.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.0 Bearer Token
security:
  - bearerAuth: []